requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cpf-generator==2.1.4
cryptography==41.0.7
defusedxml==0.7.1
deprecation==2.1.0
Django==5.0
django-allauth==0.58.2
django-betterforms==2.0.0
django-bootstrap-modal-forms==3.0.4
django-bootstrap4==23.2
django-crispy-forms==2.1
Faker==21.0.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
